Simplify detekt configuration and remove legacy versions support
