+++ plugin/src/main/groovy/com/novoda/staticanalysis/internal/detekt/DetektConfigurator.groovy
@@ -16,14 +16,12 @@ class DetektConfigurator implements Configurator {
     private static final String DETEKT_PLUGIN = 'io.gitlab.arturbosch.detekt'
     private static final String LAST_COMPATIBLE_DETEKT_VERSION = '1.0.0-RC12'
     private static final String DETEKT_NOT_APPLIED = 'The Detekt plugin is configured but not applied. Please apply the plugin in your build script.\nFor more information see https://github.com/arturbosch/detekt.'
-    private static final String OUTPUT_NOT_DEFINED = 'Output not defined! To analyze the results, `output` needs to be defined in Detekt profile.'
     private static final String DETEKT_CONFIGURATION_ERROR = "A problem occurred while configuring Detekt. Please make sure to use a compatible version (All versions up to $LAST_COMPATIBLE_DETEKT_VERSION)"
     private static final String XML_REPORT_NOT_ENABLED = 'XML report must be enabled. Please make sure to enable "reports.xml" in your Detekt configuration'
 
     private final Project project
     private final Violations violations
     private final Task evaluateViolations
-
 
     static DetektConfigurator create(Project project,
                                      NamedDomainObjectContainer<Violations> violationsContainer,
@@ -55,63 +53,36 @@ class DetektConfigurator implements Configurator {
             config.resolveStrategy = Closure.DELEGATE_FIRST
             config()
 
-            def collectViolations = configureToolTask(detekt)
+            def collectViolations = createCollectViolationsTask(violations)
             evaluateViolations.dependsOn collectViolations
         }
     }
 
     private void setDefaultXmlReport(detekt) {
-        if (detekt.hasProperty('reports')) {
+        try {
             detekt.reports {
                 xml.enabled = true
                 xml.destination = new File(project.buildDir, 'reports/detekt/detekt.xml')
             }
+        } catch (Exception exception) {
+            throw new GradleException(DETEKT_CONFIGURATION_ERROR, exception)
         }
     }
 
-    private def configureToolTask(detekt) {
-        def detektTask = project.tasks.findByName('detekt')
-        if (detektTask?.hasProperty('reports')) {
-            def reports = detektTask.reports
-            if (!reports.xml.enabled) {
-                throw new IllegalStateException(XML_REPORT_NOT_ENABLED)
-            }
-            return createCollectViolationsTask(
-                    violations,
-                    detektTask,
-                    reports.xml.destination,
-                    reports.html.destination
-            )
-        }
-
-        // Fallback to old Detekt versions
-        def output = resolveOutput(detekt)
-        if (!output) {
-            throw new IllegalArgumentException(OUTPUT_NOT_DEFINED)
-        }
-        def outputFolder = project.file(output)
-        return createCollectViolationsTask(
-                violations,
-                project.tasks['detektCheck'],
-                new File(outputFolder, 'detekt-checkstyle.xml'),
-                new File(outputFolder, 'detekt-report.html')
-        )
-    }
-
-    private static resolveOutput(detekt) {
-        if (detekt.hasProperty('profileStorage')) {
-            detekt.profileStorage.systemOrDefault.output
-        } else if (detekt.respondsTo('systemOrDefaultProfile')) {
-            detekt.systemOrDefaultProfile().output
-        } else {
-            throw new IllegalStateException(DETEKT_CONFIGURATION_ERROR)
-        }
-    }
-
-    private def createCollectViolationsTask(Violations violations, detektTask, File xmlReportFile, File htmlReportFile) {
+    private def createCollectViolationsTask(Violations violations) {
         createTask(project, 'collectDetektViolations', CollectCheckstyleViolationsTask) { task ->
-            task.xmlReportFile = xmlReportFile
-            task.htmlReportFile = htmlReportFile
+            def detektTask = project.tasks.findByName('detekt')
+            try {
+                def reports = detektTask.reports
+                if (!reports.xml.enabled) {
+                    throw new IllegalStateException(XML_REPORT_NOT_ENABLED)
+                }
+
+                task.xmlReportFile = reports.xml.destination
+                task.htmlReportFile = reports.html.destination
+            } catch (Exception exception) {
+                throw new GradleException(DETEKT_CONFIGURATION_ERROR, exception)
+            }
             task.violations = violations
             task.dependsOn detektTask
         }
